requirements.txt
pandas~=2.1.3
requests~=2.31.0
beautifulsoup4~=4.12.2
mysqlclient~=2.2.0
python-dotenv~=1.0.0
numpy~=1.26.2
SQLAlchemy~=2.0.23
langchain~=0.0.343
streamlit~=1.28.2